Use underscores in PR template placeholders
On macOS, and probably on other operating systems, double clicking a word, selects it. The OS sees underscores as word characters, which isn't the case for hyphens, so that this_word is seen as one word, whereas this-word isn't. Replacing the dashes with underscores makes these placeholders easier to select and replace.